More updates

+++ remote-control/motor-controller-mini/include/pwm_output.h
@@ -7,20 +7,20 @@ inline void pwm_init() {
 }
 
 inline void pwm_start() {
-    for (int i = 0; i < pwm_outputs_count; i++) {
+    for (short i = 0; i < pwm_outputs_count; i++) {
         ledcAttachPin(pwm_output_pins[i], i);
     }
 }
 
 inline void pwm_stop() {
-    for (int i = 0; i < pwm_outputs_count; i++) {
+    for (short i = 0; i < pwm_outputs_count; i++) {
         ledcDetachPin(pwm_output_pins[i]);
     }
 }
 
 template<short TMin, short TMax>
 inline void pwm_write(const short values[], const short size) {
-    for (int i = 0; i < min(size, pwm_outputs_count); i++) {
+    for (short i = 0; i < min(size, pwm_outputs_count); i++) {
          ledcWrite(i, map(values[i], TMin, TMax, SERVO_LOW, SERVO_HIGH));
     }
 }
